Moves sampleClients.js out of where it shouldn't be
Maybe not to a better place, though
